checks for creating profile cont.
